Fri, Jun 9, 2023, 4:01 PM +03:00

+++ hours/hours.c
@@ -46,7 +46,7 @@ float calc_hours(int hours[], int weeks, char output)
             sum += hours[i]  ;
         }
 
-        float avg = 0;
+        double avg = 0;
         for (int j = 0; j < output; j++)
         {
             avg = sum / weeks ;
